Refactor internal and public API
Adaptations to changes applied in the context of eclipse-glsp/glsp#1
